+++ dnssd/advertiser/dnsimple/doc.go
@@ -1,0 +1,3 @@
+// Package dnsimple provides a [dnssd.Advertiser] implementation that advertises
+// DNS-SD service instances on domain names hosted by dnsimple.com.
+package dnsimple