+++ CHANGELOG.md
@@ -15,7 +15,8 @@ The format is based on [Keep a Changelog], and this project adheres to
 ### Added
 
 - Added `dnssd.Advertiser` interface.
-- Added `route53.Advertiser`, which advertises DNS-SD services via AWS Route 53.
+- Added `route53.Advertiser`, which advertises DNS-SD services on domains hosted by AWS Route 53.
+- Added `dnsimple.Advertiser`, which advertises DNS-SD services on domains hosted by dnsimple.com.
 
 ### Changed